backend/lambdas/event-messaging/requirements.txt
boto3
botocore
pyjwt